pyproject.toml
[project]
name = "anki-scroll"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "erwan serandour", email = "e.serandour@hispeed.ch" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "jinja2>=3.1.6",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]

[project.scripts]
anki-scroll = "anki_scroll:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
